+++ src/app/myinfo/components/BoardItem.tsx
