+++ src/app/myinfo/components/SmallBoardMyInfo.tsx
